restructure of the directories
