+++ src/config/app_config.rs
@@ -1,0 +1,6 @@
+
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct AppConfig {
+
+}
