+++ src/config/mod.rs
@@ -1,0 +1,3 @@
+mod app_config;
+
+pub use app_config::AppConfig
