--- src/stuff/mod.rs
@@ -1,3 +1,0 @@
-pub mod library;
-pub mod player;
-pub mod playlist;
